work around cyclical dependency when setting up mgmt
At the moment, the `mgmt` environment needs the Jenkins AMI to be
available to fully build the infrastructure. The AMI, however, can't be
built without parts of the infrastructure being present. This sets up a
minimal subset of the infrastructure as a first step to allow the AMI
build to happen, then adds the rest of the infrastructure later.